requirements.txt
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.3
scikit-image>=0.18.3
scikit-learn>=0.24.2
opencv-python>=4.5.3
tensorflow>=2.8.0
keras>=2.8.0
Pillow>=8.3.2
tqdm>=4.62.3
pyyaml>=6.0
